Removed QPGrid class. Either fixed-point and grid must be derived from the same parent, or there must be no class at all (imo)! TODO: Refactor fixed-point, grid methods using the new sigma functionality (i.e. calculate individual sigma_c diag elements). Then make some kind of QP equation solver class.
